# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: JA4proxy Management UI
  description: Management interface for JA4proxy TLS security proxy
  version: 1.0.0
paths:
...
  /api/v1/lists/{list_type}/{list_name}:
    get:
      tags:
      - lists
      summary: Get List
      description: Return all entries in the specified list.
      operationId: get_list_api_v1_lists__list_type___list_name__get
      parameters:
      - name: list_type
        in: path
        required: true
        schema:
          type: string
          title: List Type
      - name: list_name
        in: path
        required: true
        schema:
          type: string
          title: List Name
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEntries'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/partials/dial:
    get:
      tags:
      - partials
      summary: Dial Partial
      description: Return the dial widget HTML fragment with the current dial value.
      operationId: dial_partial_api_v1_partials_dial_get
      parameters:
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/partials/bans:
    get:
      tags:
      - partials
      summary: Bans Partial
      description: Return the bans table HTML fragment.
      operationId: bans_partial_api_v1_partials_bans_get
      parameters:
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/partials/audit:
    get:
      tags:
      - partials
      summary: Audit Partial
      description: Return audit log table rows as an HTML fragment.
      operationId: audit_partial_api_v1_partials_audit_get
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Pagination offset
          default: 0
          title: Offset
        description: Pagination offset
      - name: filter
        in: query
        required: false
        schema:
          type: string
          description: 'Filter: all | dial | ban | list | reload'
          default: all
          title: Filter
        description: 'Filter: all | dial | ban | list | reload'
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/partials/list-table:
    get:
      tags:
      - partials
      summary: List Table Partial
      description: Return a list entries table as an HTML fragment.
      operationId: list_table_partial_api_v1_partials_list_table_get
      parameters:
      - name: list
        in: query
        required: true
        schema:
          type: string
          description: 'List name: ja4_whitelist | ja4_blacklist | ip_allowlist'
          title: List
        description: 'List name: ja4_whitelist | ja4_blacklist | ip_allowlist'
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            text/html:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ListEntries:
      properties:
        entries:
          items:
            type: string
          type: array
          title: Entries
        count:
          type: integer
          minimum: 0.0
          title: Count
      type: object
      required:
      - count
      title: ListEntries
      description: Response from GET /api/v1/lists/{list_type}/{list_name}.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...